Move to virtual doms to greatly speed up channels with high counts
