+++ package.json
@@ -1,1 +1,1 @@
-{"name":"neonplug","version":"0.1.0","type":"module","description":"Cyberpunk-themed Radio CPS for Baofeng DM-32UV","scripts":{"dev":"vite","build":"tsc && vite build","build:prod":"tsc && vite build --mode production","build:single":"tsc && vite build --mode singlefile","preview":"vite preview","lint":"eslint . --report-unused-disable-directives --max-warnings 0","test":"vitest","audit:prod":"npm audit --omit=dev"},"dependencies":{"react":"^18.2.0","react-dom":"^18.2.0","react-easy-crop":"^5.5.6","zustand":"^4.4.7"},"devDependencies":{"@eslint/js":"^9.39.5","@types/react":"^18.2.43","@types/react-dom":"^18.2.17","@vitejs/plugin-react":"^4.7.0","@vitest/coverage-v8":"^4.1.7","autoprefixer":"^10.4.16","eslint":"^9.39.5","eslint-plugin-react-hooks":"^7.1.1","eslint-plugin-react-refresh":"^0.4.5","globals":"^17.8.0","postcss":"^8.4.32","tailwindcss":"^3.4.0","typescript":"^5.2.2","typescript-eslint":"^8.65.0","vite":"^7.3.6","vite-plugin-singlefile":"^2.3.0","vitest":"^4.0.16"}}
+{"name":"neonplug","version":"0.1.0","type":"module","description":"Cyberpunk-themed Radio CPS for Baofeng DM-32UV","scripts":{"dev":"vite","build":"tsc && vite build","build:prod":"tsc && vite build --mode production","build:single":"tsc && vite build --mode singlefile","preview":"vite preview","lint":"eslint . --report-unused-disable-directives --max-warnings 0","test":"vitest","audit:prod":"npm audit --omit=dev"},"dependencies":{"@tanstack/react-virtual":"^3.14.9","react":"^18.2.0","react-dom":"^18.2.0","react-easy-crop":"^5.5.6","zustand":"^4.4.7"},"devDependencies":{"@eslint/js":"^9.39.5","@types/react":"^18.2.43","@types/react-dom":"^18.2.17","@vitejs/plugin-react":"^4.7.0","@vitest/coverage-v8":"^4.1.7","autoprefixer":"^10.4.16","eslint":"^9.39.5","eslint-plugin-react-hooks":"^7.1.1","eslint-plugin-react-refresh":"^0.4.5","globals":"^17.8.0","postcss":"^8.4.32","tailwindcss":"^3.4.0","typescript":"^5.2.2","typescript-eslint":"^8.65.0","vite":"^7.3.6","vite-plugin-singlefile":"^2.3.0","vitest":"^4.0.16"}}
